CRUD

+++ src/main/java/jakub/library/library/model/Book.java
@@ -1,24 +1,24 @@
 package jakub.library.library.model;
 
 public class Book {
-    private long id;
+    private int id;
     private String title;
     private String author;
 
     public Book() {
     }
 
-    public Book(long id, String title, String author) {
+    public Book(Integer id, String title, String author) {
         this.id = id;
         this.title = title;
         this.author = author;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
